Load: Support remote load in import-tsfile script (#13352)
(cherry picked from commit 8a2aba23c0673e0560b7aff2ffc9b3ce9a73a5ff)
