+++ iotdb-client/cli/src/main/java/org/apache/iotdb/tool/schema/AbstractSchemaTool.java
@@ -17,12 +17,10 @@
  * under the License.
  */
 
-package org.apache.iotdb.tool;
+package org.apache.iotdb.tool.schema;
 
 import org.apache.iotdb.cli.utils.IoTPrinter;
 import org.apache.iotdb.exception.ArgsErrorException;
-import org.apache.iotdb.rpc.IoTDBConnectionException;
-import org.apache.iotdb.rpc.StatementExecutionException;
 import org.apache.iotdb.session.Session;
 
 import org.apache.commons.cli.CommandLine;
@@ -37,10 +35,11 @@ import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.io.PrintWriter;
-import java.time.ZoneId;
+import java.util.Arrays;
+import java.util.Iterator;
 import java.util.List;
 
-public abstract class AbstractDataTool {
+public abstract class AbstractSchemaTool {
 
   protected static final String HOST_ARGS = "h";
   protected static final String HOST_NAME = "host";
@@ -60,56 +59,11 @@ public abstract class AbstractDataTool {
   protected static final String USERNAME_NAME = "username";
   protected static final String USERNAME_DEFAULT_VALUE = "root";
 
-  protected static final String TIME_FORMAT_ARGS = "tf";
-  protected static final String TIME_FORMAT_NAME = "timeformat";
-
-  protected static final String TIME_ZONE_ARGS = "tz";
-  protected static final String TIME_ZONE_NAME = "timeZone";
-
   protected static final String TIMEOUT_ARGS = "timeout";
-  protected static final String TIMEOUT_NAME = "timeout";
+  protected static final String TIMEOUT_ARGS_NAME = "queryTimeout";
+
   protected static final int MAX_HELP_CONSOLE_WIDTH = 92;
-  protected static final String[] TIME_FORMAT =
-      new String[] {"default", "long", "number", "timestamp"};
-  protected static final String[] STRING_TIME_FORMAT =
-      new String[] {
-        "yyyy-MM-dd HH:mm:ss.SSSX",
-        "yyyy/MM/dd HH:mm:ss.SSSX",
-        "yyyy.MM.dd HH:mm:ss.SSSX",
-        "yyyy-MM-dd HH:mm:ssX",
-        "yyyy/MM/dd HH:mm:ssX",
-        "yyyy.MM.dd HH:mm:ssX",
-        "yyyy-MM-dd HH:mm:ss.SSSz",
-        "yyyy/MM/dd HH:mm:ss.SSSz",
-        "yyyy.MM.dd HH:mm:ss.SSSz",
-        "yyyy-MM-dd HH:mm:ssz",
-        "yyyy/MM/dd HH:mm:ssz",
-        "yyyy.MM.dd HH:mm:ssz",
-        "yyyy-MM-dd HH:mm:ss.SSS",
-        "yyyy/MM/dd HH:mm:ss.SSS",
-        "yyyy.MM.dd HH:mm:ss.SSS",
-        "yyyy-MM-dd HH:mm:ss",
-        "yyyy/MM/dd HH:mm:ss",
-        "yyyy.MM.dd HH:mm:ss",
-        "yyyy-MM-dd'T'HH:mm:ss.SSSX",
-        "yyyy/MM/dd'T'HH:mm:ss.SSSX",
-        "yyyy.MM.dd'T'HH:mm:ss.SSSX",
-        "yyyy-MM-dd'T'HH:mm:ssX",
-        "yyyy/MM/dd'T'HH:mm:ssX",
-        "yyyy.MM.dd'T'HH:mm:ssX",
-        "yyyy-MM-dd'T'HH:mm:ss.SSSz",
-        "yyyy/MM/dd'T'HH:mm:ss.SSSz",
-        "yyyy.MM.dd'T'HH:mm:ss.SSSz",
-        "yyyy-MM-dd'T'HH:mm:ssz",
-        "yyyy/MM/dd'T'HH:mm:ssz",
-        "yyyy.MM.dd'T'HH:mm:ssz",
-        "yyyy-MM-dd'T'HH:mm:ss.SSS",
-        "yyyy/MM/dd'T'HH:mm:ss.SSS",
-        "yyyy.MM.dd'T'HH:mm:ss.SSS",
-        "yyyy-MM-dd'T'HH:mm:ss",
-        "yyyy/MM/dd'T'HH:mm:ss",
-        "yyyy.MM.dd'T'HH:mm:ss"
-      };
+
   protected static final int CODE_OK = 0;
   protected static final int CODE_ERROR = 1;
 
@@ -117,18 +71,16 @@ public abstract class AbstractDataTool {
   protected static String port;
   protected static String username;
   protected static String password;
-  protected static ZoneId zoneId;
 
-  protected static String timeZoneID;
-  protected static String timeFormat;
-  protected static String exportType;
   protected static String aligned;
   protected static Session session;
 
+  protected static final List<String> HEAD_COLUMNS =
+      Arrays.asList("Timeseries", "Alias", "DataType", "Encoding", "Compression");
   private static final IoTPrinter ioTPrinter = new IoTPrinter(System.out);
-  private static final Logger LOGGER = LoggerFactory.getLogger(AbstractDataTool.class);
+  private static final Logger LOGGER = LoggerFactory.getLogger(AbstractSchemaTool.class);
 
-  protected AbstractDataTool() {}
+  protected AbstractSchemaTool() {}
 
   protected static String checkRequiredArg(
       String arg, String name, CommandLine commandLine, String defaultValue)
@@ -146,36 +98,11 @@ public abstract class AbstractDataTool {
     return str;
   }
 
-  protected static void setTimeZone() throws IoTDBConnectionException, StatementExecutionException {
-    if (timeZoneID != null) {
-      session.setTimeZone(timeZoneID);
-    }
-    zoneId = ZoneId.of(session.getTimeZone());
-  }
-
   protected static void parseBasicParams(CommandLine commandLine) throws ArgsErrorException {
     host = checkRequiredArg(HOST_ARGS, HOST_NAME, commandLine, HOST_DEFAULT_VALUE);
     port = checkRequiredArg(PORT_ARGS, PORT_NAME, commandLine, PORT_DEFAULT_VALUE);
     username = checkRequiredArg(USERNAME_ARGS, USERNAME_NAME, commandLine, USERNAME_DEFAULT_VALUE);
-    password = commandLine.getOptionValue(PW_ARGS, PW_DEFAULT_VALUE);
-  }
-
-  protected static boolean checkTimeFormat() {
-    for (String format : TIME_FORMAT) {
-      if (timeFormat.equals(format)) {
-        return true;
-      }
-    }
-    for (String format : STRING_TIME_FORMAT) {
-      if (timeFormat.equals(format)) {
-        return true;
-      }
-    }
-    LOGGER.info(
-        "Input time format {} is not supported, "
-            + "please input like yyyy-MM-dd\\ HH:mm:ss.SSS or yyyy-MM-dd'T'HH:mm:ss.SSS%n",
-        timeFormat);
-    return false;
+    password = checkRequiredArg(PW_ARGS, PW_NAME, commandLine, PW_DEFAULT_VALUE);
   }
 
   protected static Options createNewOptions() {
@@ -184,6 +111,7 @@ public abstract class AbstractDataTool {
     Option opHost =
         Option.builder(HOST_ARGS)
             .longOpt(HOST_NAME)
+            .optionalArg(true)
             .argName(HOST_NAME)
             .hasArg()
             .desc("Host Name (optional)")
@@ -193,6 +121,7 @@ public abstract class AbstractDataTool {
     Option opPort =
         Option.builder(PORT_ARGS)
             .longOpt(PORT_NAME)
+            .optionalArg(true)
             .argName(PORT_NAME)
             .hasArg()
             .desc("Port (optional)")
@@ -202,6 +131,7 @@ public abstract class AbstractDataTool {
     Option opUsername =
         Option.builder(USERNAME_ARGS)
             .longOpt(USERNAME_NAME)
+            .optionalArg(true)
             .argName(USERNAME_NAME)
             .hasArg()
             .desc("Username (optional)")
@@ -223,19 +153,14 @@ public abstract class AbstractDataTool {
   /**
    * write data to CSV file.
    *
-   * @param headerNames the header names of CSV file
    * @param records the records of CSV file
    * @param filePath the directory to save the file
    */
-  public static Boolean writeCsvFile(
-      List<String> headerNames, List<List<Object>> records, String filePath) {
+  public static Boolean writeCsvFile(List<List<Object>> records, String filePath) {
     try {
       final CSVPrinterWrapper csvPrinterWrapper = new CSVPrinterWrapper(filePath);
-      if (headerNames != null) {
-        csvPrinterWrapper.printRecord(headerNames);
-      }
       for (List<Object> CsvRecord : records) {
-        csvPrinterWrapper.printRecord(CsvRecord);
+        csvPrinterWrapper.printRecordln(CsvRecord);
       }
       csvPrinterWrapper.flush();
       csvPrinterWrapper.close();
@@ -267,6 +192,19 @@ public abstract class AbstractDataTool {
         csvPrinter = csvFormat.print(new PrintWriter(filePath));
       }
       csvPrinter.printRecord(values);
+    }
+
+    public void printRecordln(final Iterable<?> values) throws IOException {
+      if (csvPrinter == null) {
+        csvPrinter = csvFormat.print(new PrintWriter(filePath));
+      }
+      Iterator var2 = values.iterator();
+
+      while (var2.hasNext()) {
+        Object value = var2.next();
+        csvPrinter.print(value);
+      }
+      csvPrinter.println();
     }
 
     public void print(Object value) {
