+++ iotdb-client/cli/src/test/java/org/apache/iotdb/tool/WriteDataFileTest.java
@@ -19,6 +19,8 @@
 
 package org.apache.iotdb.tool;
 
+import org.apache.iotdb.tool.data.AbstractDataTool;
+
 import org.junit.Before;
 import org.junit.Test;
